Limpio formulario importando funcion validadora y comento codigo para que se entienda
